feat: Added a default goal to recycle unneeded items.
This change raises the macOS deployment target to 10.15. It also introduces a `DefaultRecycleGoal` that recycles items when the total count in both inventory and bank exceeds a certain threshold. The character will move to the correct workshop and recycle the excess items. The priority of the default crafting goal has been adjusted to accommodate this new recycling goal.